still a bug when parsing hex

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: herbie <herbie@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 21:47:58 by herbie            #+#    #+#             */
-/*   Updated: 2022/11/15 19:00:45 by herbie           ###   ########.fr       */
+/*   Updated: 2022/11/16 09:59:35 by herbie           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,10 +27,10 @@ int	ft_intlen(int n);
 char	*ft_strchr(const char *s, int c);
 int	ft_putnbr_fd(int n, int fd);
 int	ft_putunbr_fd(unsigned int n, int fd);
-int	ft_puthex_fd(long n, int fd);
+int	ft_puthex_fd(unsigned long n, int fd);
 int	ft_puthex_up_fd(long long n, int fd);
-int ft_putptr_fd(unsigned long long n, int fd);
-int	ft_putnbr_base_fd(unsigned long long n, char *base, int fd);
+int ft_putptr_fd(unsigned long n, int fd);
+int	ft_putnbr_base_fd(unsigned long n, char *base, int fd);
 int	ft_putstr_fd(char *s, int fd);
 int	ft_putchar_fd(char c, int fd);
